more project structure refactoring
