+++ app/src/TodoApp/reducers/VisibilityFilter.js
